Moves OpenCL version of Reprojection module to its own namespace
