+++ module/vision/OpenCL/Reprojection/tests/Reprojection.cpp
@@ -1,0 +1,6 @@
+// Uncomment this line when other test files are added
+//#define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
+//#include <catch.hpp>
+
+// Remove this line when test files are added
+int main() { return 0; }
